BSIS-3250 Add server_uuid to key of AuditRevision
This commit adds a server uuid to the primary
key of Audit Revision. It also updates all
referencing Audit tables. It obtains the
server uuid from the mysql global variable
@@server_uuid.

The commit then adds triggers on the AuditRevision
and Audit tables to set this server_uuid value when
it is not supplied in the the INSERT values. If
the server_uuid is supplied in the list of INSERT
values it will not be set by the trigger.

The commit also updates the type of the auto-increment
id from integer to BIGINT. The commit includes the
updates in the code base.

+++ src/test/java/org/jembi/bsis/helpers/builders/AuditRevisionViewModelBuilder.java
@@ -10,12 +10,12 @@ import org.jembi.bsis.viewmodel.UserViewModel;
 
 public class AuditRevisionViewModelBuilder extends AbstractBuilder<AuditRevisionViewModel> {
 
-  private int id;
+  private long id;
   private Date revisionDate;
   private UserViewModel user;
   private Set<EntityModification> entityModifications = new HashSet<>();
 
-  public AuditRevisionViewModelBuilder withId(int id) {
+  public AuditRevisionViewModelBuilder withId(long id) {
     this.id = id;
     return this;
   }
